Move `arrayToTable` so it is available for unit tests
