+++ javascripts/discourse-table-builder/lib/utilities.js
