+++ javascripts/discourse/api-initializers/table-editor.js
@@ -9,7 +9,7 @@ import { popupAjaxError } from "discourse/lib/ajax-error";
 import Session from "discourse/models/session";
 import loadScript from "discourse/lib/load-script";
 import PrettyText, { buildOptions } from "pretty-text/pretty-text";
-import { tokenRange } from "../lib/utilities";
+import { tokenRange } from "../discourse-table-builder/lib/utilities";
 export default apiInitializer("0.11.1", (api) => {
   const site = api.container.lookup("site:main"),
     siteSettings = api.container.lookup("site-settings:main");
